requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
celery==5.5.3
redis==6.4.0
pydantic==2.11.7
python-dotenv==1.1.1
sqlalchemy==2.0.43
psycopg[binary]==3.2.10
pytest==8.4.1
httpx==0.28.1
